polymarket-bot/requirements.txt
py-clob-client==0.34.6
web3==6.14.0
python-dotenv==1.0.0
requests==2.31.0
